Added the 1Spioc logo to the header along with the Sidebar.tsx file moved to the proper Sidebar folder
