+++ src/App.tsx
@@ -21,7 +21,7 @@ import SaveIcon from "@mui/icons-material/Save";
 import AddIcon from "@mui/icons-material/Add";
 import DeleteIcon from "@mui/icons-material/Close";
 import MoreVertIcon from "@mui/icons-material/MoreVert";
-import Sidebar from "./components/Sidebar";
+import Sidebar from "./components/Sidebar/Sidebar";
 import Builder from "./components/Builder";
 import MobilePreview from "./components/MobilePreview";
 import MetaJsonGenerator from "./components/MetaJsonGenerator";
